Fix circular gamestate dependency
