--- src/game/state/GameState.js
@@ -1,5 +1,0 @@
-export default class GameState {
-  constructor(name) {
-    this.name = name;
-  }
-}
