+++ src/game/state/GameState.ts
@@ -1,0 +1,11 @@
+export abstract class GameState {
+  readonly name: string;
+
+  constructor(name: string) {
+    this.name = name;
+  }
+
+  abstract init(): void;
+  abstract receiveInputText(text: string): void;
+  abstract redraw(): void;
+}
